cleanup. fixed console errors
