+++ app/client/src/App.js
@@ -7,7 +7,7 @@ import {
 } from 'react-router-dom';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import { connect } from 'react-redux';
-import { MuiThemeProvider } from '@material-ui/core/styles';
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import MainLayout from './layouts/MainLayout';
 import EmptyLayout from './layouts/EmptyLayout';
 
@@ -32,6 +32,8 @@ import Register from './containers/Register/Register';
 import UserNotice from './components/UserNotice/UserNotice';
 
 firebase.initializeApp(config.firebaseConfig);
+
+const theme = createMuiTheme({});
 
 const LOGIN_STATE = {
     SUCCESS: 'SUCCESS',
@@ -95,7 +97,7 @@ function App(props) {
     return (
         <>
             {user.loggedIn === LOGIN_STATE.SUCCESS ? (
-                <MuiThemeProvider>
+                <MuiThemeProvider theme={theme}>
                     <CssBaseline />
                     <div style={{ height: '100vh' }}>
                         <Router>
